Nearley performance benchmark using a JSON grammar and 1,000 lines input.
it is very slow, too slow to add to the benchmark with multiple Parsing solutions...

This is also not the complete JSON grammar.
Not whitespace handling (as it seems to slow nearley even more)
and without complete JSON string escapes as that too slows Nearley
way too much.

This is meant to reproduce the low performance with a minimal example.

Not added t