Fix: change angle brackets to backticks in comment
This change was done b/c the angle brackets break the compiler for [typedoc-plugin-markdown](https://github.com/tgreyuk/typedoc-plugin-markdown), a library for converting Typedoc to markdown. It interprets the angle brackets as an HTML tag, whereas it's just being used for emphasis.

This issue occurred specifically when using [docusaurus-plugin-typedoc](https://www.npmjs.com/package/docusaurus-plugin-typedoc). More details about this problem can be found here - https://github.com/tgreyuk/typedoc-plugin-markdown/issues/276

This change fixes the issues by replacing the backticks (`<` and `>`) with markdown backticks to highlight the word and achieve the same desired emphasis without breaking the markdown parser.

i was able to validate that this works in my local environment.